[pr-3] Move SDK error types to embedding-sdk-shared
