+++ enterprise/frontend/src/metabase-enterprise/embedding_iframe_sdk/auth-manager/AuthManager.ts
@@ -7,7 +7,7 @@ import {
 import {
   INVALID_AUTH_METHOD,
   MetabaseError,
-} from "embedding-sdk-bundle/errors";
+} from "embedding-sdk-shared/errors";
 import type { EmbedAuthManagerContext } from "metabase/embedding/embedding-iframe-sdk/types/auth-manager";
 import type { MetabaseEmbeddingSessionToken } from "metabase/embedding-sdk/types/refresh-token";
