+++ frontend/src/embedding-sdk-shared/errors/base.ts
