+++ frontend/src/embedding-sdk-shared/errors/error-code.ts
