+++ frontend/src/embedding-sdk-shared/errors/error-docs-links.ts
