+++ frontend/src/embedding-sdk-shared/errors/generic.ts
