+++ frontend/src/embedding-sdk-shared/errors/index.ts
