+++ frontend/src/embedding-sdk-shared/errors/jwt.ts
