+++ frontend/src/embedding-sdk-shared/errors/saml.ts
